Standardizing Mobile<->Wear messaging
