+++ mobile/src/main/java/dev/jhale/android/wear/simpletimetracker/MainActivity.kt
@@ -6,16 +6,15 @@
 package dev.jhale.android.wear.simpletimetracker
 
 import android.annotation.SuppressLint
-import android.content.BroadcastReceiver
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.widget.Button
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
 
-
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     lateinit var textView: TextView
     lateinit var buttonDebug: Button
@@ -26,7 +25,6 @@ class MainActivity : AppCompatActivity() {
     @SuppressLint("MissingInflatedId")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val messageSTTBroadcastTransmitter = STTBroadcastTransmitter()
 
         setContentView(R.layout.activity_main)
         this.textView = findViewById(R.id.text_view_message)
@@ -34,11 +32,6 @@ class MainActivity : AppCompatActivity() {
 
 
         this.buttonDebug = findViewById(R.id.button_debug)
-        this.buttonDebug.setOnClickListener {
-            messageSTTBroadcastTransmitter.onReceive(applicationContext, Intent().also{ intent ->
-                intent.putExtra("message", "Personal Learning|Hobbies")
-            })
-        }
 
         this.buttonDebugCategories = findViewById(R.id.button_debug_read_categories)
         this.buttonDebugCategories.setOnClickListener {
@@ -64,41 +57,6 @@ class MainActivity : AppCompatActivity() {
                 it.setPackage("com.razeeman.util.simpletimetracker.debug")
                 Log.i(LOG_TAG, "Broadcasting Intent: $it")
                 applicationContext.sendBroadcast(it)
-            }
-        }
-    }
-
-    inner class STTBroadcastTransmitter() : BroadcastReceiver() {
-        override fun onReceive(context: Context?, intent: Intent?) {
-            val action = intent?.action
-
-            when (action) {
-                //This branch not used... see STTBroadcastReceiver
-                "com.razeeman.util.simpletimetracker.ACTION_SEND_CATEGORIES" -> {
-                    intent.getStringExtra("categories").also {
-                        categoriesView.text = it
-                    }
-                }
-                else -> {
-                    val message = intent?.extras?.getString("message") ?: "|"
-                    textView.text = message
-
-                    val messageParts = message.split("|")
-                    val activity = messageParts[0]
-                    val tag = messageParts[1]
-                    if (activity.isNotEmpty()) {
-                        Intent().also { intent ->
-                            intent.action = "com.razeeman.util.simpletimetracker.ACTION_START_ACTIVITY"
-                            intent.setPackage("com.razeeman.util.simpletimetracker.debug")
-                            intent.putExtra("extra_activity_name", activity)
-                            if (tag.isNotEmpty()) {
-                                intent.putExtra("extra_record_tag", tag)
-                            }
-                            Log.i(LOG_TAG, "Broadcasting Intent: $intent")
-                            context?.sendBroadcast(intent)
-                        }
-                    }
-                }
             }
         }
     }
