+++ mobile/src/main/java/dev/jhale/android/wear/simpletimetracker/STTBroadcastReceiver.kt
@@ -5,6 +5,7 @@ import android.content.Context
 import android.content.Intent
 import android.util.Log
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 
 @AndroidEntryPoint
 class STTBroadcastReceiver : BroadcastReceiver() {
