+++ wear/src/main/res/values/wear.xml
@@ -1,0 +1,8 @@
+<resources xmlns:tools="http://schemas.android.com/tools"
+    tools:keep="@array/android_wear_capabilities">
+    <string-array name="android_wear_capabilities" tools:ignore="ExtraTranslation,UnusedResources" translatable="false">
+        <item>receive_categories</item>
+        <item>receive_running_records</item>
+        <item>receive_prefs</item>
+    </string-array>
+</resources>
